+++ R/api.R
@@ -1,17 +1,39 @@
-callJS <- function(method) {
+callJS <- function() {
   # get the parameters from the function that have a value
   message <- Filter(function(x) !is.symbol(x), as.list(parent.frame(1)))
   session <- shiny::getDefaultReactiveDomain()
 
-  # call the JS function
-  method <- paste0("timevis:", method)
-  session$sendCustomMessage(method, message)
+  # If a timevis widget was passed in, this is during a chain pipeline in the
+  # initialization of the widget, so keep track of the desired function call
+  # by adding it to a list of functions that should be performed when the widget
+  # is ready
+  if (methods::is(message$id, "timevis")) {
+    widget <- message$id
+    message$id <- NULL
+    widget$x$api
+    widget$x$api <- c(widget$x$api, list(message))
+    return(widget)
+  }
+  # If an ID was passed, the widget already exists and we can simply call the
+  # appropriate JS function
+  else if (is.character(message$id)) {
+    method <- paste0("timevis:", message$method)
+    session$sendCustomMessage(method, message)
+    return(message$id)
+  } else {
+    stop("The `id` argument must be either a timevis htmlwidget or an ID
+         of a timevis htmlwidget.", call. = FALSE)
+  }
 }
 
 #' Add a single item to a timeline
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param data A named list containing the item data to add.
 #' @examples
+#'
+#' timevis() %>%
+#'   addItem(list(start = Sys.Date(), content = "Today"))
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -32,13 +54,18 @@ callJS <- function(method) {
 #' @export
 addItem <- function(id, data) {
   method <- "addItem"
-  callJS(method)
+  callJS()
 }
 
 #' Add multiple items to a timeline
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param data A dataframe containing the items data to add.
 #' @examples
+#'
+#' timevis() %>%
+#'   addItems(data.frame(start = c(Sys.Date(), Sys.Date() - 1),
+#'            content = c("Today", "Yesterday")))
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -62,13 +89,17 @@ addItem <- function(id, data) {
 addItems <- function(id, data) {
   method <- "addItems"
   data <- dataframeToD3(data)
-  callJS(method)
+  callJS()
 }
 
 #' Remove an item from a timeline
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param itemId The id of the item to remove
 #' @examples
+#'
+#' timevis(data.frame(id = 1:2, start = Sys.Date(), content = c("1", "2"))) %>%
+#'   removeItem(2)
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -91,14 +122,18 @@ addItems <- function(id, data) {
 #' @export
 removeItem <- function(id, itemId) {
   method <- "removeItem"
-  callJS(method)
+  callJS()
 }
 
-#' Add new vertical bar at a time point that can be dragged by the user
-#' @param id Timeline id
+#' Add a new vertical bar at a time point that can be dragged by the user
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param time The date/time to add
 #' @param itemId The id of the custom time bar
 #' @examples
+#'
+#' timevis() %>%
+#'   addCustomTime(Sys.Date() - 1, "yesterday")
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -119,13 +154,19 @@ removeItem <- function(id, itemId) {
 #' @export
 addCustomTime <- function(id, time, itemId) {
   method <- "addCustomTime"
-  callJS(method)
+  callJS()
 }
 
 #' Remove a custom time previously added
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param itemId The id of the custom time bar
 #' @examples
+#'
+#' timevis() %>%
+#'   addCustomTime(Sys.Date() - 1, "yesterday") %>%
+#'   addCustomTime(Sys.Date() + 1, "tomorrow") %>%
+#'   removeCustomTime("yesterday")
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -150,11 +191,11 @@ addCustomTime <- function(id, time, itemId) {
 #' @export
 removeCustomTime <- function(id, itemId) {
   method <- "removeCustomTime"
-  callJS(method)
+  callJS()
 }
 
 #' Adjust the visible window such that it fits all items
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param options Named list of options controlling the animation. Most common
 #' option is \code{"animation" = TRUE/FALSE}. For a full list of options, see
 #' the "fit" method in the
@@ -183,11 +224,11 @@ removeCustomTime <- function(id, itemId) {
 #' @export
 fitWindow <- function(id, options) {
   method <- "fitWindow"
-  callJS(method)
+  callJS()
 }
 
 #' Move the window such that the given time is centered
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param time The date/time to center around
 #' @param options Named list of options controlling the animation. Most common
 #' option is \code{"animation" = TRUE/FALSE}. For a full list of options, see
@@ -195,6 +236,10 @@ fitWindow <- function(id, options) {
 #' \href{http://visjs.org/docs/timeline/#Methods}{official
 #' Timeline documentation}
 #' @examples
+#'
+#' timevis() %>%
+#'   centerTime(Sys.Date() - 1)
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -207,7 +252,7 @@ fitWindow <- function(id, options) {
 #'       timevis()
 #'     )
 #'     observeEvent(input$btn, {
-#'       centerTime("timeline", Sys.Date())
+#'       centerTime("timeline", Sys.Date() - 1)
 #'     })
 #'   }
 #' )
@@ -215,11 +260,11 @@ fitWindow <- function(id, options) {
 #' @export
 centerTime <- function(id, time, options) {
   method <- "centerTime"
-  callJS(method)
+  callJS()
 }
 
 #' Move the window such that given item or items are centered
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param itemId A vector (or single value) of the item ids to center
 #' @param options Named list of options controlling mainly the animation.
 #' Most common option is \code{"animation" = TRUE/FALSE}. For a full list of
@@ -227,6 +272,14 @@ centerTime <- function(id, time, options) {
 #' \href{http://visjs.org/docs/timeline/#Methods}{official
 #' Timeline documentation}
 #' @examples
+#'
+#' timevis(data.frame(
+#'           id = 1:3,
+#'           start = c(Sys.Date() - 1, Sys.Date(), Sys.Date() + 1),
+#'           content = c("Item 1", "Item 2", "Item 3"))
+#' ) %>%
+#'   centerItem(1)
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -251,13 +304,17 @@ centerTime <- function(id, time, options) {
 #' @export
 centerItem <- function(id, itemId, options) {
   method <- "centerItem"
-  callJS(method)
+  callJS()
 }
 
 #' Set the items of a timeline
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param data A dataframe containing the item data to use.
 #' @examples
+#'
+#' timevis(data.frame(start = Sys.Date(), content = "Today")) %>%
+#'   setItems(data.frame(start = Sys.Date() - 1, content = "yesterday"))
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -280,15 +337,66 @@ centerItem <- function(id, itemId, options) {
 setItems <- function(id, data) {
   method <- "setItems"
   data <- dataframeToD3(data)
-  callJS(method)
+  callJS()
+}
+
+#' Set the groups of a timeline
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
+#' @param data A dataframe containing the groups data to use.
+#' @examples
+#'
+#' timevis(data = data.frame(
+#'   start = c(Sys.Date(), Sys.Date(), Sys.Date() + 1, Sys.Date() + 2),
+#'   content = c("one", "two", "three", "four"),
+#'   group = c(1, 2, 1, 2)),
+#'   groups = data.frame(id = 1:2, content = c("G1", "G2"))
+#' ) %>%
+#'   setGroups(data.frame(id = 1:2, content = c("Group 1", "Group 2")))
+#'
+#'
+#' if (interactive()) {
+#' library(shiny)
+#' shinyApp(
+#'   ui = fluidPage(
+#'     timevisOutput("timeline"),
+#'     actionButton("btn", "Change group names")
+#'   ),
+#'   server = function(input, output) {
+#'     output$timeline <- renderTimevis(
+#'       timevis(data = data.frame(
+#'                start = c(Sys.Date(), Sys.Date(), Sys.Date() + 1, Sys.Date() + 2),
+#'                          content = c("one", "two", "three", "four"),
+#'                          group = c(1, 2, 1, 2)),
+#'               groups = data.frame(id = 1:2, content = c("G1", "G2")))
+#'
+#'     )
+#'     observeEvent(input$btn, {
+#'       setGroups("timeline",
+#'                data.frame(id = 1:2, content = c("Group 1", "Group 2")))
+#'     })
+#'   }
+#' )
+#' }
+#' @export
+setGroups <- function(id, data) {
+  method <- "setGroups"
+  data <- dataframeToD3(data)
+  callJS()
 }
 
 #' Update the configuration options of a timeline
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param options A named list containing updated configuration options to use.
 #' See the \code{options} parameter of the
 #' \code{\link[timevis]{timevis}} function to see more details.
 #' @examples
+#'
+#' timevis(
+#'   data.frame(start = Sys.Date(), content = "Today"),
+#'   options = list(showCurrentTime = FALSE, orientation = "top")
+#' ) %>%
+#'   setOptions(list(editable = TRUE, showCurrentTime = TRUE))
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -312,11 +420,11 @@ setItems <- function(id, data) {
 #' @export
 setOptions <- function(id, options) {
   method <- "setOptions"
-  callJS(method)
+  callJS()
 }
 
 #' Select one or multiple items on a timeline
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param itemId A vector (or single value) of the item ids to select
 #' @param options Named list of options controlling mainly the animation.
 #' Most common options are \code{focus = TRUE/FALSE} and
@@ -325,6 +433,10 @@ setOptions <- function(id, options) {
 #' \href{http://visjs.org/docs/timeline/#Methods}{official
 #' Timeline documentation}
 #' @examples
+#'
+#' timevis(data.frame(id = 1:3, start = Sys.Date(), content = 1:3)) %>%
+#'   setSelection(2)
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -347,11 +459,11 @@ setOptions <- function(id, options) {
 #' @export
 setSelection <- function(id, itemId, options) {
   method <- "setSelection"
-  callJS(method)
+  callJS()
 }
 
 #' Set the current visible window
-#' @param id Timeline id
+#' @param id Timeline id or a \code{timevis} object (the output from \code{timevis()})
 #' @param start The start date/time to show in the timeline
 #' @param end The end date/time to show in the timeline
 #' @param options Named list of options controlling mainly the animation.
@@ -360,6 +472,10 @@ setSelection <- function(id, itemId, options) {
 #' \href{http://visjs.org/docs/timeline/#Methods}{official
 #' Timeline documentation}
 #' @examples
+#'
+#' timevis() %>%
+#'   setWindow(Sys.Date() - 1, Sys.Date() + 1)
+#'
 #' if (interactive()) {
 #' library(shiny)
 #' shinyApp(
@@ -380,5 +496,5 @@ setSelection <- function(id, itemId, options) {
 #' @export
 setWindow <- function(id, start, end, options) {
   method <- "setWindow"
-  callJS(method)
+  callJS()
 }