+++ R/utils.R
@@ -14,8 +14,16 @@ is.bool <- function(x) {
 
 # Convert a data.frame to a list of lists (the data format that D3 uses)
 dataframeToD3 <- function(df) {
-  if (missing(df)) {
+  if (missing(df) || is.null(df)) {
     return(list())
   }
+  if (!is.data.frame(df)) {
+    stop("timevis: the input must be a dataframe", call. = FALSE)
+  }
+  row.names(df) <- NULL
   apply(df, 1, function(row) as.list(row[!is.na(row)]))
 }
+
+#' @importFrom magrittr %>%
+#' @export
+magrittr::`%>%`