+++ inst/example/ui.R
@@ -1,7 +1,36 @@
 library(timevis)
 
+source("ui-helpers.R")
+
 fluidPage(
   title = "timevis - An R package for creating timeline visualizations",
+  tags$head(
+    tags$link(href = "style.css", rel = "stylesheet"),
+
+    # Favicon
+    tags$link(rel = "shortcut icon", type="image/x-icon", href="http://daattali.com/shiny/img/favicon.ico"),
+
+    # Facebook OpenGraph tags
+    tags$meta(property = "og:title", content = share$title),
+    tags$meta(property = "og:type", content = "website"),
+    tags$meta(property = "og:url", content = share$url),
+    tags$meta(property = "og:image", content = share$image),
+    tags$meta(property = "og:description", content = share$description),
+
+    # Twitter summary cards
+    tags$meta(name = "twitter:card", content = "summary"),
+    tags$meta(name = "twitter:site", content = paste0("@", share$twitter_user)),
+    tags$meta(name = "twitter:creator", content = paste0("@", share$twitter_user)),
+    tags$meta(name = "twitter:title", content = share$title),
+    tags$meta(name = "twitter:description", content = share$description),
+    tags$meta(name = "twitter:image", content = share$image)
+  ),
+  tags$a(
+    href="https://github.com/daattali/timevis",
+    tags$img(style="position: absolute; top: 0; right: 0; border: 0;",
+             src="github-orange-right.png",
+             alt="Fork me on GitHub")
+  ),
   div(id = "header",
     div(id = "title",
       "timevis"
@@ -12,35 +41,73 @@ fluidPage(
         "By",
         tags$a(href = "http://deanattali.com/", "Dean Attali"),
         HTML("&bull;"),
-        "Package & examples",
+        "Available",
         tags$a(href = "https://github.com/daattali/timevis", "on GitHub"),
         HTML("&bull;"),
         tags$a(href = "http://daattali.com/shiny/", "More apps"), "by Dean"
     )
   ),
   tabsetPanel(
+    id = "mainnav",
     tabPanel(
       div(icon("calendar"), "Basic timeline"),
-      br(),
-      tags$link(href = "style.css", rel = "stylesheet"),
-      timevisOutput("timelineBasic")
+      timevisOutput("timelineBasic"),
+      div(
+        id = "samplecode",
+        fluidRow(
+          column(
+            6,
+            div(class = "codeformat",
+              "In R console or R markdown documents"),
+            tags$pre(codeConsole)
+          ),
+          column(
+            6,
+            div(class = "codeformat",
+                "In Shiny apps"),
+            tags$pre(codeShiny)
+          )
+        )
+      ),
+      div(class = "sourcecode",
+        "The exact code for all the timelines in this app is",
+        tags$a(href = "https://github.com/daattali/timevis/tree/master/inst/example",
+               "on GitHub")
+      )
     ),
 
     tabPanel(
-      div(icon("cog"), "Custom style and parameters"),
-      br(),
-      timevisOutput("timelineCustom")
+      div(icon("cog"), "Custom style"),
+      timevisOutput("timelineCustom"),
+      div(class = "sourcecode",
+          "The exact code for all the timelines in this app is",
+          tags$a(href = "https://github.com/daattali/timevis/tree/master/inst/example",
+                 "on GitHub")
+      )
     ),
 
     tabPanel(
       div(icon("trophy"), "World Cup 2014"),
-      br(),
-      timevisOutput("timelineWC")
+      timevisOutput("timelineWC"),
+      div(class = "sourcecode",
+          "The exact code for all the timelines in this app is",
+          tags$a(href = "https://github.com/daattali/timevis/tree/master/inst/example",
+                 "on GitHub")
+      )
+    ),
+
+    tabPanel(
+      div(icon("users"), "Groups"),
+      timevisOutput("timelineGroups"),
+      div(class = "sourcecode",
+          "The exact code for all the timelines in this app is",
+          tags$a(href = "https://github.com/daattali/timevis/tree/master/inst/example",
+                 "on GitHub")
+      )
     ),
 
     tabPanel(
       div(icon("sliders"), "Fully interactive"),
-      br(),
       fluidRow(
         column(
           8,
@@ -60,8 +127,7 @@ fluidPage(
                   actionButton("focus2", "Focus item 4"),
                   actionButton("focusSelection", "Focus current selection"),
                   actionButton("addTime", "Add a draggable vertical bar 2016-01-17")
-              ),
-              br()
+              )
             )
           ),
           fluidRow(
@@ -103,6 +169,11 @@ fluidPage(
                  textOutput("selected", inline = TRUE))
            )
         )
+      ),
+      div(class = "sourcecode",
+          "The exact code for all the timelines in this app is",
+          tags$a(href = "https://github.com/daattali/timevis/tree/master/inst/example",
+                 "on GitHub")
       )
     )
   )